Add random number generator and color functions
This commit adds the necessary code changes to implement a random number generator and color functions. The files "azar.h" and "azar.c" contain the functions for generating random numbers and initializing the random seed. The files "colours.h" and "colours.c" contain the functions for changing the terminal color. These changes will be used in future code development.